[agent] fix(gaze-types): move remaining JSON serde tests to audit crate
Step 3 of task: fix PR #180 blockers
